rename components and extract check input
