+++ web/html/src/components/input/index.ts
@@ -1,4 +1,5 @@
 export * from "./check/Check";
+export * from "./check/CheckInput";
 export * from "./check/DEPRECATED_Check";
 export * from "./datetime/DateTime";
 export * from "./form/Form";
