+++ web/html/src/manager/admin/setup/products/product-check/ProductCheck.example.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 
 import { Button } from "components/buttons";
 
-import { ProductCheck } from "./product-check";
+import { ProductCheck } from "./ProductCheck";
 import { type ProductSelectionState } from "./product-selection.utils";
 
 export default () => {
